make executable compile, but not run (dynamic framework cannot be loaded)
